# openapi.yaml
openapi: 3.1.0
info:
  title: Sistema Integral de Gestión de Consorcios
  description: |
    API REST para la gestión integral de consorcios. Permite administrar expensas,
    gastos del consorcio, personal y liquidaciones, tareas y presupuestos,
    comunicación interna, reserva de amenities y configuración global. El acceso a
    los recursos está restringido por roles: Administración, Departamentos
    (Inquilinos/Propietarios), Representantes y Super-Admin.
    Fase 3.5 (jun-2026): cuenta corriente por departamento — pagos se asignan FIFO desde movimientos.

    **Multitenant SaaS (Fase 7A, jul-2026)**: la plataforma es multi-tenant con
    jerarquía Super-Admin → Administración → Consorcio → Departamento. Todo
    endpoint operacional (fuera de `/auth/*` y `/me/*`) exige el header
    `X-Consorcio-Id: <id>` para resolver el consorcio activo. El backend valida
    que el usuario del JWT tenga acceso al consorcio pedido; caso contrario
    devuelve **403**. Códigos específicos: `403 cambio_password_requerido`
    (usuario con `must_change_password=True`), `403 administracion_suspendida`
    (login sobre administración inactiva), `403 sin acceso a este consorcio`
    (spoof cross-tenant), `400 X-Consorcio-Id requerido/invalido`.
  version: 1.1.0
servers:
  - url: http://localhost:8000
    description: Servidor local de desarrollo
...
security:
  - bearerAuth: []
paths:
...
  /empleados:
    get:
      tags: [Personal]
      summary: Listar empleados (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: activo
          required: false
          schema: { type: boolean, default: true }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/EmpleadoOut' } }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/AccesoDenegado' }
    post:
      tags: [Personal]
      summary: Crear empleado
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmpleadoCrear' }
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmpleadoOut' }
        '400': { $ref: '#/components/responses/PedidoInvalido' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/AccesoDenegado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '409':
          description: CUIL duplicado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /empleados/{empleado_id}:
    parameters:
      - in: path
        name: empleado_id
        required: true
        schema: { type: integer }
    get:
      tags: [Personal]
      summary: Obtener empleado
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmpleadoOut' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/AccesoDenegado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
    patch:
      tags: [Personal]
      summary: Editar empleado
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmpleadoActualizar' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmpleadoOut' }
        '400': { $ref: '#/components/responses/PedidoInvalido' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/AccesoDenegado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
    delete:
      tags: [Personal]
      summary: Eliminar empleado (soft si tiene liquidaciones)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Soft-delete
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmpleadoOut' }
        '204': { description: Hard-delete }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/AccesoDenegado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Autenticación mediante token JWT en el header Authorization.
...
  responses:
    PedidoInvalido:
      description: El pedido es inválido o le faltan campos requeridos.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NoAutenticado:
      description: No se proporcionó token de autenticación o es inválido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    AccesoDenegado:
      description: El rol del usuario no tiene permisos para acceder al recurso.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NoEncontrado:
      description: El recurso solicitado no existe.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Descripción legible del error.
          examples: ["No tiene permisos para acceder a este recurso."]

    # ---------------------------- Consorcios (Plan D) ---------------------
...
    EmpleadoOut:
      type: object
      required: [id, nombre_completo, cuil, categoria, fecha_ingreso, sueldo_basico, proveedor_id, activo]
      properties:
        id: { type: integer }
        nombre_completo: { type: string }
        cuil: { type: string }
        categoria:
          type: string
          enum: [encargado_permanente_con_vivienda, encargado_permanente_sin_vivienda, encargado_suplente, ayudante]
        fecha_ingreso: { type: string, format: date }
        fecha_egreso: { type: string, format: date, nullable: true }
        sueldo_basico: { type: number, exclusiveMinimum: 0 }
        proveedor_id: { type: integer }
        activo: { type: boolean }
    EmpleadoCrear:
      type: object
      required: [nombre_completo, cuil, categoria, fecha_ingreso, sueldo_basico, proveedor_id]
      properties:
        nombre_completo: { type: string, minLength: 1, maxLength: 255 }
        cuil: { type: string, pattern: '^\d{2}-\d{8}-\d{1}$' }
        categoria:
          type: string
          enum: [encargado_permanente_con_vivienda, encargado_permanente_sin_vivienda, encargado_suplente, ayudante]
        fecha_ingreso: { type: string, format: date }
        fecha_egreso: { type: string, format: date, nullable: true }
        sueldo_basico: { type: number, exclusiveMinimum: 0 }
        proveedor_id: { type: integer, minimum: 1 }
    EmpleadoActualizar:
      type: object
      properties:
        nombre_completo: { type: string, minLength: 1, maxLength: 255, nullable: true }
        categoria:
          type: string
          enum: [encargado_permanente_con_vivienda, encargado_permanente_sin_vivienda, encargado_suplente, ayudante]
          nullable: true
        fecha_ingreso: { type: string, format: date, nullable: true }
        fecha_egreso: { type: string, format: date, nullable: true }
        sueldo_basico: { type: number, exclusiveMinimum: 0, nullable: true }
        proveedor_id: { type: integer, minimum: 1, nullable: true }
        activo: { type: boolean, nullable: true }
...